chore: Move Factors to lib/analytics.ts
