+++ config/consts.ts
@@ -26,20 +26,11 @@ const RiskColorMap: {
   },
 };
 
-const Factors = [
-  'risk-score',
-  'flood-hazard',
-  'vulnerability',
-  'government-response',
-  'exposure',
-];
-
 const learnMoreLink = `https://drive.google.com/file/d/1TerjJtQrWxziKJ1E5BvfvRELtWJWRw2M/view?usp=drive_link`;
 const documentationLink = `https://ids-drr.readthedocs.io/en/latest/architecture/overview.html`;
 
 export {
   RiskColorMap,
-  Factors,
   learnMoreLink,
   documentationLink,
 };
