+++ lib/analytics.ts
@@ -1,0 +1,9 @@
+// Slugs the analytics dashboard treats as score-style (1-5) indicators,
+// rather than raw metric values. Sourced from the Sendai framework.
+export const Factors = [
+  'risk-score',
+  'flood-hazard',
+  'vulnerability',
+  'government-response',
+  'exposure',
+];
